refactor: 댓글 관련 api comment.ts로 이동
