+++ app/components/post/comment/comment/CommentManageButtons.tsx
@@ -1,4 +1,4 @@
-import { deleteComment } from "@/app/modules/api";
+import { deleteComment } from "@/app/modules/api/comment";
 import styles from "../@common/ManageButtons.module.css";
 import { Comment } from "@/app/modules/types";
