+++ app/components/post/comment/like/LikeComment.tsx
@@ -1,10 +1,11 @@
-import { checkUserLoginStatus, likeCommentToggle } from "@/app/modules/api";
+import { checkUserLoginStatus } from "@/app/modules/api";
 import { useEffect, useState } from "react";
 import { useAtomValue } from "jotai";
 import { userIdAtom } from "@/app/modules/atoms";
 import { Comment } from "@/app/modules/types";
 import { LoginAlert } from "../@common/LoginAlert";
 import { LikeIcon } from "./LikeIcon";
+import { likeCommentToggle } from "@/app/modules/api/comment";
 
 interface LikeCommentProps {
   comment: Comment;
